requirements.txt
flask
Gunicorn
python-dotenv
beautifulsoup4
secure-smtplib


